+++ llama_index/multi_modal_llms/__init__.py
@@ -2,7 +2,12 @@ from llama_index.multi_modal_llms.base import (
     MultiModalLLM,
     MultiModalLLMMetadata,
 )
+from llama_index.multi_modal_llms.dashscope import (
+    DashScopeMultiModal,
+    DashScopeMultiModalModels,
+)
 from llama_index.multi_modal_llms.gemini import GeminiMultiModal
+from llama_index.multi_modal_llms.ollama import OllamaMultiModal
 from llama_index.multi_modal_llms.openai import OpenAIMultiModal
 from llama_index.multi_modal_llms.replicate_multi_modal import ReplicateMultiModal
 
@@ -12,4 +17,7 @@ __all__ = [
     "MultiModalLLM",
     "OpenAIMultiModal",
     "GeminiMultiModal",
+    "DashScopeMultiModal",
+    "DashScopeMultiModalModels",
+    "OllamaMultiModal",
 ]