+++ llama_index/node_parser/relational/base_element.py
@@ -1,9 +1,11 @@
+import asyncio
 from abc import abstractmethod
 from typing import Any, Dict, List, Optional, Sequence, Tuple, cast
 
 import pandas as pd
 from tqdm import tqdm
 
+from llama_index.async_utils import DEFAULT_NUM_WORKERS, run_jobs
 from llama_index.bridge.pydantic import BaseModel, Field, ValidationError
 from llama_index.callbacks.base import CallbackManager
 from llama_index.core.response.schema import PydanticResponse
@@ -14,7 +16,9 @@ from llama_index.schema import BaseNode, Document, IndexNode, TextNode
 from llama_index.utils import get_tqdm_iterable
 
 DEFAULT_SUMMARY_QUERY_STR = """\
-What is this table about? Give a very concise summary (imagine you are adding a caption), \
+What is this table about? Give a very concise summary (imagine you are adding a new caption and summary for this table), \
+and output the real/existing table title/caption if context provided.\
+and output the real/existing table id if context provided.\
 and also output whether or not the table should be kept.\
 """
 
@@ -37,6 +41,8 @@ class TableOutput(BaseModel):
     """Output from analyzing a table."""
 
     summary: str
+    table_title: Optional[str] = None
+    table_id: Optional[str] = None
     columns: List[TableColumnOutput]
 
 
@@ -71,6 +77,12 @@ class BaseElementNodeParser(NodeParser):
         default=DEFAULT_SUMMARY_QUERY_STR,
         description="Query string to use for summarization.",
     )
+    num_workers: int = Field(
+        default=DEFAULT_NUM_WORKERS,
+        description="Num of works for async jobs.",
+    )
+
+    show_progress: bool = Field(default=True, description="Whether to show progress.")
 
     @classmethod
     def class_name(cls) -> str:
@@ -131,22 +143,49 @@ class BaseElementNodeParser(NodeParser):
         llm = cast(LLM, llm)
 
         service_context = ServiceContext.from_defaults(llm=llm, embed_model=None)
-        for element in tqdm(elements):
+
+        table_context_list = []
+        for idx, element in tqdm(enumerate(elements)):
             if element.type != "table":
                 continue
+            table_context = str(element.element)
+            if idx > 0 and str(elements[idx - 1].element).lower().strip().startswith(
+                "table"
+            ):
+                table_context = str(elements[idx - 1].element) + "\n" + table_context
+            if idx < len(elements) + 1 and str(
+                elements[idx - 1].element
+            ).lower().strip().startswith("table"):
+                table_context += "\n" + str(elements[idx + 1].element)
+
+            table_context_list.append(table_context)
+
+        async def _get_table_output(table_context: str, summary_query_str: str) -> Any:
             index = SummaryIndex.from_documents(
-                [Document(text=str(element.element))], service_context=service_context
+                [Document(text=table_context)], service_context=service_context
             )
             query_engine = index.as_query_engine(output_cls=TableOutput)
             try:
-                response = query_engine.query(self.summary_query_str)
-                element.table_output = cast(PydanticResponse, response).response
+                response = await query_engine.aquery(summary_query_str)
+                return cast(PydanticResponse, response).response
             except ValidationError:
                 # There was a pydantic validation error, so we will run with text completion
                 # fill in the summary and leave other fields blank
                 query_engine = index.as_query_engine()
-                response_txt = str(query_engine.query(self.summary_query_str))
-                element.table_output = TableOutput(summary=response_txt, columns=[])
+                response_txt = await query_engine.aquery(summary_query_str)
+                return TableOutput(summary=str(response_txt), columns=[])
+
+        summary_jobs = [
+            _get_table_output(table_context, self.summary_query_str)
+            for table_context in table_context_list
+        ]
+        summary_outputs = asyncio.run(
+            run_jobs(
+                summary_jobs, show_progress=self.show_progress, workers=self.num_workers
+            )
+        )
+        for element, summary_output in zip(elements, summary_outputs):
+            element.table_output = summary_output
 
     def get_base_nodes_and_mappings(
         self, nodes: List[BaseNode]
@@ -209,7 +248,6 @@ class BaseElementNodeParser(NodeParser):
 
         nodes = []
         cur_text_el_buffer: List[str] = []
-
         for element in elements:
             if element.type == "table":
                 # flush text buffer
@@ -224,20 +262,56 @@ class BaseElementNodeParser(NodeParser):
                 table_df = cast(pd.DataFrame, element.table)
                 table_id = element.id + "_table"
                 table_ref_id = element.id + "_table_ref"
-                # TODO: figure out what to do with columns
-                # NOTE: right now they're excluded from embedding
+
                 col_schema = "\n\n".join([str(col) for col in table_output.columns])
+
+                # We build a summary of the table containing the extracted summary, and a description of the columns
+                table_summary = str(table_output.summary)
+                if table_output.table_title:
+                    table_summary += ",\nwith the following table title:\n"
+                    table_summary += str(table_output.table_title)
+
+                table_summary += ",\nwith the following columns:\n"
+
+                for col in table_output.columns:
+                    table_summary += f"- {col.col_name}: {col.summary}\n"
+
                 index_node = IndexNode(
-                    text=str(table_output.summary),
+                    text=table_summary,
                     metadata={"col_schema": col_schema},
                     excluded_embed_metadata_keys=["col_schema"],
                     id_=table_ref_id,
                     index_id=table_id,
                 )
-                table_str = table_df.to_string()
+
+                # We serialize the table as markdown as it allow better accuracy
+                # We do not use the table_df.to_markdown() method as it generate
+                # a table with a token hngry format.
+                table_md = "|"
+                for col_name, col in table_df.items():
+                    table_md += f"{col_name}|"
+                table_md += "\n|"
+                for col_name, col in table_df.items():
+                    table_md += f"---|"
+                table_md += "\n"
+                for row in table_df.itertuples():
+                    table_md += "|"
+                    for col in row[1:]:
+                        table_md += f"{col}|"
+                    table_md += "\n"
+
+                table_str = table_summary + "\n" + table_md
                 text_node = TextNode(
                     text=table_str,
                     id_=table_id,
+                    metadata={
+                        # serialize the table as a dictionary string
+                        "table_df": str(table_df.to_dict()),
+                        # add table summary for retrieval purposes
+                        "table_summary": table_summary,
+                    },
+                    excluded_embed_metadata_keys=["table_df", "table_summary"],
+                    excluded_llm_metadata_keys=["table_df", "table_summary"],
                 )
                 nodes.extend([index_node, text_node])
             else:
@@ -250,4 +324,5 @@ class BaseElementNodeParser(NodeParser):
             nodes.extend(cur_text_nodes)
             cur_text_el_buffer = []
 
-        return nodes
+        # remove empty nodes
+        return [node for node in nodes if len(node.text) > 0]