+++ llama_index/vector_stores/registry.py
@@ -19,6 +19,7 @@ from llama_index.vector_stores.redis import RedisVectorStore
 from llama_index.vector_stores.rocksetdb import RocksetVectorStore
 from llama_index.vector_stores.simple import SimpleVectorStore
 from llama_index.vector_stores.supabase import SupabaseVectorStore
+from llama_index.vector_stores.txtai import TxtaiVectorStore
 from llama_index.vector_stores.types import VectorStore
 from llama_index.vector_stores.weaviate import WeaviateVectorStore
 
@@ -31,6 +32,7 @@ class VectorStoreType(str, Enum):
     PINECONE = "pinecone"
     OPENSEARCH = "opensearch"
     FAISS = "faiss"
+    TXTAI = "txtai"
     CASSANDRA = "cassandra"
     CHROMA = "chroma"
     CHATGPT_PLUGIN = "chatgpt_plugin"
@@ -56,6 +58,7 @@ VECTOR_STORE_TYPE_TO_VECTOR_STORE_CLASS: Dict[VectorStoreType, Type[VectorStore]
     VectorStoreType.PINECONE: PineconeVectorStore,
     VectorStoreType.OPENSEARCH: OpensearchVectorStore,
     VectorStoreType.FAISS: FaissVectorStore,
+    VectorStoreType.TXTAI: TxtaiVectorStore,
     VectorStoreType.CASSANDRA: CassandraVectorStore,
     VectorStoreType.CHROMA: ChromaVectorStore,
     VectorStoreType.CHATGPT_PLUGIN: ChatGPTRetrievalPluginClient,