+++ CHANGELOG.md
@@ -1,11 +1,132 @@
 # ChangeLog
 
-## Unreleased
+## [0.9.43] - 2024-02-03
+
+### New Features
+
+- Add multi-modal ollama (#10434)
+
+### Bug Fixes / Nits
+
+- update base class for astradb (#10435)
+
+## [0.9.42.post1] - 2024-02-02
+
+### New Features
+
+- Add Async support for Base nodes parser (#10418)
+
+## [0.9.42] - 2024-02-02
+
+### New Features
+
+- Add support for `gpt-3.5-turbo-0125` (#10412)
+- Added `create-llama` support to rag cli (#10405)
+
+### Bug Fixes / Nits
+
+- Fixed minor bugs in lance-db vector store (#10404)
+- Fixed streaming bug in ollama (#10407)
+
+## [0.9.41] - 2024-02-01
+
+### New Features
+
+- Nomic Embedding (#10388)
+- Dashvector support sparse vector (#10386)
+- Table QA with MarkDownParser and Benchmarking (#10382)
+- Simple web page reader (#10395)
+
+### Bug Fixes / Nits
+
+- fix full node content in KeywordExtractor (#10398)
+
+## [0.9.40] - 2024-01-30
+
+### New Features
+
+- Improve and fix bugs for MarkdownElementNodeParser (#10340)
+- Fixed and improve Perplexity support for new models (#10319)
+- Ensure system_prompt is passed to Perplexity LLM (#10326)
+- Extended BaseRetrievalEvaluator to include an optional PostProcessor (#10321)
+
+## [0.9.39] - 2024-01-26
+
+### New Features
+
+- Support for new GPT Turbo Models (#10291)
+- Support Multiple docs for Sentence Transformer Fine tuning(#10297)
+
+### Bug Fixes / Nits
+
+- Marvin imports fixed (#9864)
+
+## [0.9.38] - 2024-01-25
+
+### New Features
+
+- Support for new OpenAI v3 embedding models (#10279)
+
+### Bug Fixes / Nits
+
+- Extra checks on sparse embeddings for qdrant (#10275)
+
+## [0.9.37] - 2024-01-24
+
+### New Features
+
+- Added a RAG CLI utility (#10193)
+- Added a textai vector store (#10240)
+- Added a Postgresql based docstore and index store (#10233)
+- specify tool spec in tool specs (#10263)
+
+### Bug Fixes / Nits
+
+- Fixed serialization error in ollama chat (#10230)
+- Added missing fields to `SentenceTransformerRerank` (#10225)
+- Fixed title extraction (#10209, #10226)
+- nit: make chainable output parser more exposed in library/docs (#10262)
+- :bug: summary index not carrying over excluded metadata keys (#10259)
+
+## [0.9.36] - 2024-01-23
+
+### New Features
+
+- Added support for `SageMakerEmbedding` (#10207)
+
+### Bug Fixes / Nits
+
+- Fix duplicated `file_id` on openai assistant (#10223)
+- Fix circular dependencies for programs (#10222)
+- Run `TitleExtractor` on groups of nodes from the same parent document (#10209)
+- Improve vectara auto-retrieval (#10195)
+
+## [0.9.35] - 2024-01-22
 
 ### New Features
 
 - `beautifulsoup4` dependency to new optional extra `html` (#10156)
 - make `BaseNode.hash` an `@property` (#10163)
+- Neutrino (#10150)
+- feat: JSONalyze Query Engine (#10067)
+- [wip] add custom hybrid retriever notebook (#10164)
+- add from_collection method to ChromaVectorStore class (#10167)
+- CLI experiment v0: ask (#10168)
+- make react agent prompts more editable (#10154)
+- Add agent query pipeline (#10180)
+
+### Bug Fixes / Nits
+
+- Update supabase vecs metadata filter function to support multiple fields (#10133)
+- Bugfix/code improvement for LanceDB integration (#10144)
+- `beautifulsoup4` optional dependency (#10156)
+- Fix qdrant aquery hybrid search (#10159)
+- make hash a @property (#10163)
+- fix: bug on poetry install of llama-index[postgres] (#10171)
+- [doc] update jaguar vector store documentation (#10179)
+- Remove use of not-launched finish_message (#10188)
+- Updates to Lantern vector stores docs (#10192)
+- fix typo in multi_document_agents.ipynb (#10196)
 
 ## [0.9.34] - 2024-01-19